refactor : User 관련 enum 의 name 등 가독성 개선
