+++ src/main/java/com/pullanner/web/controller/oauth2/dto/OAuthAttributes.java
@@ -1,6 +1,6 @@
 package com.pullanner.web.controller.oauth2.dto;
 
-import com.pullanner.domain.user.Role;
+import com.pullanner.domain.user.enums.UserRole;
 import com.pullanner.domain.user.User;
 import com.pullanner.exception.oauth2.UnsupportedOAuth2ProviderException;
 import java.util.Map;
@@ -86,7 +86,7 @@ public class OAuthAttributes {
             .email(email)
             .provider(provider)
             .profileImageUrl(picture)
-            .role(Role.USER)
+            .userRole(UserRole.USER)
             .experiencePoint(10)
             .build();
     }
